controller-Attribut in OthelloGUI ausgelagert, GUI übergibt sich selbst jetzt an seinen BoardView
