+++ othello/view/OthelloGUI.java
@@ -8,16 +8,19 @@ import othello.controller.OthelloController;
 @SuppressWarnings("serial")
 public class OthelloGUI extends JFrame {
 
-    private final BoardView board;
+    private OthelloController controller;
+    private BoardView board;
 
     public OthelloGUI(OthelloController controller) {
 	super("Othello");
+
+	setController(controller);
 
 	setDefaultCloseOperation(EXIT_ON_CLOSE);
 
 	Occupation[][] fieldViewOccupations = new Occupation[8][8];
 	Occupation.fill(fieldViewOccupations, Occupation.NONE);
-	board = new BoardView(fieldViewOccupations, controller);
+	board = new BoardView(this, fieldViewOccupations);
 	add(board);
 
 	pack();
@@ -25,7 +28,19 @@ public class OthelloGUI extends JFrame {
 	setVisible(true);
     }
 
+    public OthelloController getController() {
+	return controller;
+    }
+
+    public void setController(OthelloController controller) {
+	this.controller = controller;
+    }
+
     public BoardView getBoard() {
 	return board;
+    }
+
+    public void setBoard(BoardView board) {
+	this.board = board;
     }
 }
